refactor[#17] : Checkbox cva 적용
